+++ src/base/SubscriptionGateway.php
@@ -531,6 +531,19 @@ abstract class SubscriptionGateway extends Gateway
     {
         $paymentIntent = $data['data']['object'];
         if ($paymentIntent['object'] === 'payment_intent') {
+
+            // Before we treat this as a normal payment intent, lets see if it was checkout session payment intent
+            $sessions = $this->getStripeClient()->checkout->sessions->all([
+                'payment_intent' => $paymentIntent['id'],
+            ]);
+
+            if($session = $sessions->data[0] ?? null)
+                $transaction = Plugin::getInstance()->getTransactions()->getTransactionByReference($session['id']);
+                $error = '';
+                Plugin::getInstance()->getPayments()->completePayment($transaction, $error);
+                return;
+            }
+
             $transaction = Plugin::getInstance()->getTransactions()->getTransactionByReference($paymentIntent['id']);
 
             if (!$transaction?->id) {